[CHORE] Moved components into Settings Page folder
