+++ scrybe_frontend/src/pages/Settings/Components/SettingsPageRedirectNav/SettingsPageRedirectNav.module.scss
